WIP EKF and supporting functions
I am finally trying to finish this and write this because I want
to start using this on a few project so I will be trying to grind
this out.

The goal of this project is to have something that *should* be easy
to debug while also have some codebase that tries to maintain a
readable flow where you can see how the data flows.

This is NOT meant to be the fastest. If I get this working
well and it doesn't perform well I will try to make it faster.
This is meant to be run on embedded devices so memory is a factor.